+++ frontend/src/components/pages/HomePage/LogRecordsTable.tsx
@@ -29,7 +29,9 @@ import EditLog from "../../forms/EditLog";
 import LogRecordAPIClient from "../../../APIClients/LogRecordAPIClient";
 import ResidentAPIClient from "../../../APIClients/ResidentAPIClient";
 import UserAPIClient from "../../../APIClients/UserAPIClient";
+import BuildingAPIClient from "../../../APIClients/BuildingAPIClient";
 import { UserLabel } from "../../../types/UserTypes";
+import { BuildingLabel } from "../../../types/BuildingTypes";
 import ConfirmationModal from "../../common/ConfirmationModal";
 
 type Props = {
@@ -56,6 +58,8 @@ const LogRecordsTable = ({
   const { authenticatedUser, setAuthenticatedUser } = useContext(AuthContext);
 
   const [showAlert, setShowAlert] = useState(false);
+
+  const [buildingOptions, setBuildingOptions] = useState<BuildingLabel[]>([]);
 
   // Menu states
   const [deleteOpenMap, setDeleteOpenMap] = useState<{
@@ -100,6 +104,15 @@ const LogRecordsTable = ({
       setResidentOptions(residentLabels);
     }
 
+    const buildingsData = await BuildingAPIClient.getBuildings();
+
+    if (buildingsData && buildingsData.buildings.length !== 0) {
+      const buildingLabels: BuildingLabel[] = buildingsData.buildings.map(
+        (building) => ({ label: building.name!, value: building.id! }),
+      );
+      setBuildingOptions(buildingLabels);
+    }
+
     const usersData = await UserAPIClient.getUsers({ returnAll: true });
     if (usersData && usersData.users.length !== 0) {
       const userLabels: UserLabel[] = usersData.users
@@ -118,11 +131,8 @@ const LogRecordsTable = ({
     } catch (error) {
       return;
     }
-    const newUserPageNum = (
-      logRecords.length === 1
-        ? userPageNum - 1 
-        : userPageNum
-    );
+    const newUserPageNum =
+      logRecords.length === 1 ? userPageNum - 1 : userPageNum;
     countRecords();
     setShowAlert(true);
     setUserPageNum(newUserPageNum);
@@ -223,6 +233,7 @@ const LogRecordsTable = ({
                       getRecords={getRecords}
                       countRecords={countRecords}
                       setUserPageNum={setUserPageNum}
+                      buildingOptions={buildingOptions}
                     />
 
                     <ConfirmationModal
